genesis app_state 独立出文件
